[BUGFIX] Check availability of cache before access
getPageTsConfig now checks if cache manager has a cache defined
which is not the case in all situations where the method was called
(e.g. during installation).

This change prevents an exception which causes improper (partially
aborted) installation on a fresh TYPO3 instance.

Unit test
onlyFetchRootTypoScriptOfRootlineIfTheFluxConfigurationManagerIsInjected
has been split, renamed and fixed (was not testing what the name
suggested).

Fixes #355