[BUGFIX] Avoid MiscellaneousUtility for icons, force relative path
This change fixes by hardcoding the expected relative path of an icon and avoids calling Flux's MIscellaneousUtility to generate icons (since this Utility is not and probably cannot be fully SVG-compatible on TYPO3 6.2 without IconRegistry). Icons are shown in full size and must be manually resized if a different size is desired.

Does not affect methods of defining the icon. Convention path based resolving and Flux form options still work the same way.