[TASK] Remove INFO logging of built content setups
This change removes a TYPO3 syslog INFO notice sent when pageTSconfig gets built for a page. The information was intended as a way for administrators to track problems if content setup cannot be built (the error is logged the same way and this is NOT removed).

Close: https://github.com/FluidTYPO3/flux/issues/1157