[REMOVAL] Mark extension obsolete in favor of Flux
Flux now provides the same features without the need for a subtype. For new projects, Flux can be used vanilla to create content types - and for old projects that don't wish to migrate or just need more time to do so, this final major version of fluidcontent can still be installed.